Improvements (#36)
* use the template button and improve the custom style

* Added a theme switch and animated the theme switch

* improved the sun rays

* use the correct layout

* Fixed installation of package and components

* Added Examples to test the designs

* added rounded button

* Next and previous are now roundbuttons

* used alias for namespaced targets when add subdirectory

* do not build examples if not instructed

* Added the close button

* Added switch

* Added textarea and text field

* Added tooltip

* Added popup

* Removed old interface

* Fixed the readmes

* use the common cmake to build docs from qml
