+++ .github/workflows/build-test-install.yml
@@ -14,14 +14,15 @@ jobs:
         os: [ubuntu-latest,macos-latest,windows-latest]
         sharedLib: [true]
 
-    uses: EddyTheCo/Common/.github/workflows/build-test-install.yml@main
+    uses: EddyTheCo/Common/.github/workflows/build-test-install.yml@v0.1.1
     permissions:
       contents: write
     with:
       os: ${{ matrix.os }}
       sharedLib: ${{ matrix.sharedLib }}
       qtModules: 'qtshadertools' 
-      qtVersion: '6.6.0'  
+      qtVersion: '6.6.0' 
+      test: false  
   release:
     if: startsWith(github.ref, 'refs/tags/v')
     needs: build_test_package
