+++ .gitignore
@@ -1,0 +1,6 @@
+.git/
+doc/html/
+doc/*.tag
+*.swp
+*.autosave
+*.wasm
