+++ CMakeLists.txt
@@ -1,17 +1,17 @@
 cmake_minimum_required(VERSION 3.16)
-
+include(${CMAKE_CURRENT_BINARY_DIR}/local_conf.cmake OPTIONAL)
 include(FetchContent)
 FetchContent_Declare(
 	ccommon
 	GIT_REPOSITORY https://github.com/EddyTheCo/Common.git
-	GIT_TAG main
+	GIT_TAG v0.1.1
 	)
 FetchContent_MakeAvailable(ccommon)
 version_from_git(
 	LOG       OFF
 	TIMESTAMP "%Y%m%d%H%M%S"
 	)
-project(MyDesigns VERSION ${VERSION} LANGUAGES CXX)
+project(EstervDesigns VERSION ${VERSION} LANGUAGES CXX)
 
 set_property(GLOBAL PROPERTY USE_FOLDERS ON)
 set(CMAKE_CXX_STANDARD 17)
@@ -21,6 +21,7 @@ set(CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS ON)
 include(GNUInstallDirs)
 
 find_package(Qt6 COMPONENTS Core Gui Quick Qml  OPTIONAL_COMPONENTS  ShaderTools )
+
 if (Qt6_FOUND)
 	qt_standard_project_setup()
 	if(NOT TARGET Qt6::ShaderTools)
@@ -31,100 +32,48 @@ if (Qt6_FOUND)
 			)
 		FetchContent_MakeAvailable(qtshadertools)
 	endif()
-	set_source_files_properties(qml/CustomStyle.qml PROPERTIES
-		QT_QML_SINGLETON_TYPE TRUE
-		)
-	qt6_add_qml_module(MyDesigns
-		URI  MyDesigns
-		VERSION 1.0
-		SOURCES
-		src/mydesigns.cpp include/mydesigns.hpp
-		QML_FILES
-		qml/MyTextField.qml
-		qml/AmountText.qml
-		qml/MyTextArea.qml
-		qml/MyFrame.qml
-		qml/MyButton.qml
-		qml/MyLabel.qml
-		qml/MyRadioButton.qml
-		qml/MyCheckBox.qml
-		qml/MySettButton.qml
-		qml/Notification.qml
-		qml/PlusButton.qml
-		qml/CustomStyle.qml
-		qml/CloseButton.qml	
-		RESOURCE_PREFIX
-		"/esterVtech.com/imports"
-		OUTPUT_TARGETS out_targets_var
-		OUTPUT_DIRECTORY
-		${CMAKE_BINARY_DIR}/MyDesigns
-		IMPORT_PATH ${CMAKE_BINARY_DIR}
-		)
 
-	target_include_directories(MyDesigns PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-		"$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/MyDesigns>")
-
-
-	qt6_add_shaders(MyDesigns "esterVtech.com.imports.MyDesigns.shaders"
-		BATCHABLE
-		PRECOMPILE
-		OPTIMIZED
-		OUTPUT_TARGETS out_targets_var2
-		PREFIX
-		"/esterVtech.com/imports/MyDesigns"
-		FILES
-		"frag/mySettButton.frag"
-		"frag/arrowHead.frag"
-		"frag/plus.frag"
-		"frag/gausianBlur.frag"
-		"frag/voronoi.frag"
-		"frag/close.frag"
-                "frag/qrscanner.frag"
-		)
+	add_subdirectory(Styles)
+	add_subdirectory(Controls)
+	add_subdirectory(CustomControls)
 
 
 
-	target_link_libraries(MyDesigns PUBLIC Qt6::Gui Qt6::Quick)
-
-
-	install(TARGETS MyDesigns MyDesignsplugin ${out_targets_var}  ${out_targets_var2}
-		EXPORT MyDesignsTargets
-		DESTINATION ${CMAKE_INSTALL_LIBDIR}
-		)
-
-	install(DIRECTORY ${CMAKE_CURRENT_LIST_DIR}/include/
-		DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/MyDesigns
-		)
-	install(DIRECTORY ${CMAKE_BINARY_DIR}/MyDesigns/
-		DESTINATION ${CMAKE_INSTALL_LIBDIR}/QMLPlugins/MyDesigns
-		)
-	install(EXPORT MyDesignsTargets
-		FILE MyDesignsTargets.cmake
-		DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/MyDesigns
+	install(EXPORT ${PROJECT_NAME}-config
+		FILE ${PROJECT_NAME}-config.cmake
+		DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/Esterv
+		NAMESPACE ${PROJECT_NAME}::
+		COMPONENT All
 		)
 	include(CMakePackageConfigHelpers)
 	configure_package_config_file(${CMAKE_CURRENT_SOURCE_DIR}/Config.cmake.in
 		"${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}Config.cmake"
-		INSTALL_DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/MyDesigns
+		INSTALL_DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/Esterv
 		)
 	write_basic_package_version_file(
-		"${CMAKE_CURRENT_BINARY_DIR}/MyDesignsConfigVersion.cmake"
+		"${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}ConfigVersion.cmake"
 		VERSION ${VERSION}
 		COMPATIBILITY SameMajorVersion
 		)
 	install(FILES
-		${CMAKE_CURRENT_BINARY_DIR}/MyDesignsConfig.cmake
-		${CMAKE_CURRENT_BINARY_DIR}/MyDesignsConfigVersion.cmake
-		DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/MyDesigns
+		${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}Config.cmake
+		${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}ConfigVersion.cmake
+		DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/Esterv
 		)
-	export(EXPORT MyDesignsTargets
-		FILE "${CMAKE_CURRENT_BINARY_DIR}/MyDesignsTargets.cmake"
+	export(EXPORT ${PROJECT_NAME}-config
+		FILE "${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}-config.cmake"
 		)
 	if(CMAKE_PROJECT_NAME STREQUAL PROJECT_NAME)
+		add_subdirectory(examples)
 		set(CPACK_PACKAGE_CONTACT "estervtech")
-		include(CTest)
 		set(CPACK_PACKAGE_FILE_NAME "${PROJECT_NAME}-v${SEMVER}-${CMAKE_SYSTEM_NAME}_${CMAKE_SYSTEM_VERSION}-${CMAKE_SYSTEM_PROCESSOR}-${CMAKE_CXX_COMPILER_ID}")
 		include(CPack)
 	endif()
-endif(Qt6_FOUND)
+else()
+	message(STATUS "${PROJECT_NAME}" will not be build)
+endif()
 
+if(BUILD_DOCS)
+	get_target_property(build_docs cmake_build_docs SOURCES)
+	include(${build_docs})
+endif()
