+++ Config.cmake.in
@@ -1,4 +1,4 @@
 @PACKAGE_INIT@
 include(CMakeFindDependencyMacro)
 find_dependency(Qt6 REQUIRED COMPONENTS Gui Quick )
-include ( "${CMAKE_CURRENT_LIST_DIR}/MyDesignsTargets.cmake" )
+include ( "${CMAKE_CURRENT_LIST_DIR}/EstervDesigns-config.cmake" )
