+++ CustomControls/CMakeLists.txt
@@ -1,0 +1,38 @@
+
+qt6_add_qml_module(customControls
+	URI  Esterv.CustomControls
+	VERSION 1.0
+	QML_FILES
+        "qml/PrevButton.qml"
+        "qml/NextButton.qml"
+        "qml/ThemeSwitch.qml"
+        "qml/CloseButton.qml"
+	RESOURCE_PREFIX
+	"/esterVtech.com/imports"
+	OUTPUT_TARGETS out_targets_var
+	OUTPUT_DIRECTORY
+	${CMAKE_BINARY_DIR}/Esterv/CustomControls
+	IMPORT_PATH ${CMAKE_BINARY_DIR}
+	)
+
+add_library(${PROJECT_NAME}::customControls ALIAS customControls)
+add_library(${PROJECT_NAME}::customControlsplugin ALIAS customControlsplugin)
+
+target_link_libraries(customControls PUBLIC Qt6::Gui Qt6::Quick SimpleStyle )
+
+
+install(TARGETS customControls ${out_targets_var}
+        EXPORT ${PROJECT_NAME}-config
+	DESTINATION ${CMAKE_INSTALL_LIBDIR}
+	COMPONENT CustomControls
+	)
+install(DIRECTORY ${CMAKE_BINARY_DIR}/Esterv/CustomControls
+	DESTINATION ${CMAKE_INSTALL_LIBDIR}/Esterv
+	COMPONENT CustomControls
+	)
+install(TARGETS customControlsplugin 
+        EXPORT ${PROJECT_NAME}-config
+	DESTINATION ${CMAKE_INSTALL_LIBDIR}/Esterv/CustomControls
+	COMPONENT CustomControls
+	)
+
