--- examples/window.qml
@@ -1,16 +1,0 @@
-import QtQuick.Controls
-import QtQuick
-import MyDesigns
-import CustomStyle 1.0
-
-ApplicationWindow {
-    visible: true
-    width:1200
-    height:700
-
-    background: Rectangle {
-        color:"#1e1e1e"
-    }
-
-
-}
