--- qml/MyLabel.qml
@@ -1,5 +1,0 @@
-import QtQuick.Controls
-
-Label {
-    color: CustomStyle.midColor2
-}
