--- src/mydesigns.cpp
@@ -1,1 +1,0 @@
-#include "mydesigns.hpp"
